refact: move error declaration ot another module
