+++ src/parser/error/mod.rs
@@ -1,0 +1,3 @@
+mod parse_error;
+
+pub use parse_error::{ErrorCode, ParseError};
