+++ src/parser/mod.rs
@@ -1,9 +1,9 @@
+mod error;
 mod facade;
 mod fs;
 mod multi_module;
-mod parse_error;
 mod single_module;
 
+pub use error::*;
 pub use facade::Parser;
 pub use fs::{FileSystem, FileSystemError};
-pub use parse_error::{ErrorCode, ParseError};
