+++ src/parser/single_module/interface.rs
@@ -1,7 +1,7 @@
 use std::path::Path;
 
 use super::super::super::ast::ModuleTerm;
-use super::super::parse_error::ParseError;
+use super::super::error::ParseError;
 
 pub trait SingleModuleParser {
     fn parse(&self, path: &Path) -> Result<Box<ModuleTerm>, ParseError>;
